It works now and some errors less

+++ Client RLD/src/Map.cpp
@@ -18,7 +18,7 @@ void Map::loadFromFile(std::string path, std::vector<std::shared_ptr<sf::Texture
 	int j = 0;
 	if (file.is_open()) {
 		while (std::getline(file, line)) {
-			for (int i = 0; i < line.size(); i += 2) {
+			for (size_t i = 0; i < line.size(); i += 2) {
 				tileArray[i / 2][j] = std::make_shared<Tile>((TileType)(line[i] - 'A'), *tex[line[i] - 'A']);
 			}
 			j++;
